front-end/test/authtest/requirements.txt
streamlit==1.45.0
requests
datetime
st-annotated-text
streamlit-timeline

